// src/api/model/friendshipDetailResult.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */

export interface FriendshipDetailResult {
  friendId?: number;
  friendNickname?: string;
  friendProfileImageUrl?: string;
  friendAlias?: string;
  intimacy?: number;
  myInterestScore?: number;
  isMuted?: boolean;
  isRoutable?: boolean;
}